feat: Add docker-compose file for tei-grpc-rerank service
This commit adds a new docker-compose file, `docker-compose-grpc.yaml`, which defines the configuration for the `tei-grpc-rerank` service. The service uses the `ghcr.io/huggingface/text-embeddings-inference:cpu-latest-grpc` image and exposes port `50083`. It also mounts the `./data` directory as a volume and sets the `--model-id` command line argument to `BAAI/bge-reranker-large`. This docker-compose file enables the deployment and execution of the `tei-grpc-rerank` service with the specified configuration.